Refactor DashboardChartComponent to remove unused imports and update file name
