+++ Common/Types/Dashboard/DashboardComponents/DashboardChartComponent.ts
